pyproject.toml
[project]
name = "woobot2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-community>=0.3.22",
    "langchain-core>=0.3.56",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "langgraph-checkpoint-redis>=0.0.4",
    "ngrok>=1.4.0",
    "python-dotenv>=1.1.0",
    "rapidfuzz>=3.13.0",
    "uvicorn>=0.34.2",
    "woocommerce>=3.0.0",
]
